convert indx to long

+++ pcdet/models/dense_heads/detr/matcher.py
@@ -59,7 +59,7 @@ class HungarianMatcher(nn.Module):
         out_bbox = outputs["box_preds"].flatten(0, 1)  # [batch_size * num_queries, 4]
 
         # Also concat the target labels and boxes
-        tgt_ids = torch.cat([v["labels"] for v in targets])
+        tgt_ids = torch.cat([v["labels"] for v in targets]).long()
         tgt_bbox = torch.cat([v["boxes"] for v in targets])
 
         # Compute the classification cost. Contrary to the loss, we don't use the NLL,
